+++ src/openweather_sdk/rest/forecast.py
@@ -6,8 +6,9 @@ from openweather_sdk.validators import _validate_selected_attr
 class _ForecastAPI:
     """
     A class for creating data for path buildng to Forecast API.
-    See: https://openweathermap.org/forecast5 or
-    https://openweathermap.org/api/hourly-forecast.
+    See: https://openweathermap.org/forecast5,
+    https://openweathermap.org/api/hourly-forecast or
+    https://openweathermap.org/forecast16.
     """
 
     def __init__(self, lon, lat, appid, **kwargs):
@@ -27,7 +28,7 @@ class _ForecastAPI:
     def version(self, value):
         self._version = _validate_selected_attr(value, _FORECAST_API_VERSIONS)
 
-    def _get_forecast_5_day(self):
+    def _get_forecast_5_days(self):
         """Get 5 day forecast (3-hour step) at the specified point."""
         end_point = "forecast"
         query_params = {
@@ -43,6 +44,19 @@ class _ForecastAPI:
     def _get_forecast_hourly(self):
         """Get hourly forecast for 4 days (96 timestamps) at the specified point."""
         end_point = "forecast/hourly"
+        query_params = {
+            "lat": self.lat,
+            "lon": self.lon,
+            "appid": self.appid,
+            "units": self.units,
+            "lang": self.language,
+        }
+        url = _build_url(self.service_name, self.version, end_point, query_params)
+        return _APIRequest(url)._get_data()
+
+    def _get_forecast_daily_16_days(self):
+        """Get daily forecast for 16 days at the specified point."""
+        end_point = "forecast/daily"
         query_params = {
             "lat": self.lat,
             "lon": self.lon,